% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.wknn.mi.R
\docType{package}
\name{wkNNMI}
\alias{wkNNMI}
\title{wkNNMI: An Adaptive Mutual Information-Weighted k-NN Algorithm for the
Imputation of Static and Dynamic Mixed-Type Data}
\description{
This package implements an adaptive weighted k-nearest neighbours (wk-NN)
imputation algorithm for clinical register data developed to explicitly
handle missing values of continuous/ordinal/categorical and static/dynamic
features conjointly. For each subject with missing data to be imputed,
the method creates a feature vector constituted by the information collected
over his/her first *window_size* time units of visits. This vector is used as
sample in a k-nearest neighbours procedure, in order to select, among the
other patients, the ones with the most similar temporal evolution of the
disease over time. An *ad hoc* similarity metric was implemented for
the sample comparison, capable of handling the different nature of the data,
the presence of multiple missing values and include the cross-information
among features.
}
\details{
The wkNNMI package mainly serves as container for the two functions that
implement the imputation algorithm impute.subject() and impute.wknn(), and
for the example datasets patient.data and new.patient.
}
